Update fn.php - PHP 8 compatibility
Since version 8 of PHP, the syntax `$string == true` always returns false. This means that even if everything is set up correctly, it will not work. Therefore, it is necessary to replace this check with a comparison against an empty string. This should work well with earlier versions of PHP as well.